rename webviz -> wikiviz
