+++ backend/scrapy/scrapy.cfg
@@ -1,5 +1,5 @@
 [settings]
-default = webvis.settings
+default = wikivis.settings
 
 [deploy]
-project = webvis
+project = wikivis
