+++ backend/scrapy/wikivis/utils/network/exporter.py
@@ -1,7 +1,7 @@
 from networkx import Graph
 from pyvis.network import Network
 
-from webvis.utils.create_missing_folders import create_missing_folders
+from wikivis.utils.create_missing_folders import create_missing_folders
 
 
 class Exporter:
